battlefield_service.go -> delete obsolete types and functions battle_service.go -> add IsFull() from deleted battlefield_service.go
