+++ service/message_service.go
@@ -35,12 +35,6 @@ func ProcessWebhookMessage(update *telegram.Update) {
 	}
 }
 
-func sendHintIfUnavailable(callbackQuery *telegram.CallbackQuery, emoji string) {
-	if strings.HasPrefix(callbackQuery.Data, "PRESS_UNAVAILABLE") {
-		SendAnswerCallbackQuery(callbackQuery.ID, "☹️You can capture neighboring cells only:\n" + "↖️🔼↗️\n◀️" + emoji + "▶️\n↙️🔽↘️", true)
-	}
-}
-
 func SendMessage(chatID int64, message string, replyMarkup *telegram.InlineKeyboardMarkup) {
 	err := sendMessage(chatID, message, replyMarkup)
 	if err != nil {
@@ -70,6 +64,12 @@ func convertToString(update *telegram.Update) string {
 	}
 
 	return string(out)
+}
+
+func sendHintIfUnavailable(callbackQuery *telegram.CallbackQuery, emoji string) {
+	if strings.HasPrefix(callbackQuery.Data, "PRESS_UNAVAILABLE") {
+		SendAnswerCallbackQuery(callbackQuery.ID, "☹️You can capture neighboring cells only:\n" + "↖️🔼↗️\n◀️" + emoji + "▶️\n↙️🔽↘️", true)
+	}
 }
 
 func sendMessage(chatID int64, message string, replyMarkup *telegram.InlineKeyboardMarkup) error {
